--- a/include/sonic/internal/arch/neon/skip.h
+++ b/include/sonic/internal/arch/neon/skip.h
@@ -18,8 +18,9 @@
 
 #define VEC_LEN 16
 
-#include <sonic/internal/utils.h>
 #include <sonic/internal/arch/common/skip_common.h>
+#include <sonic/internal/utils.h>
+
 #include "base.h"
 #include "simd.h"
 
@@ -34,7 +35,7 @@ using sonic_json::internal::common::SkipLiteral;
 
 sonic_force_inline bool SkipContainer(const uint8_t *data, size_t &pos,
                                       size_t len, uint8_t left, uint8_t right) {
-    return skip_container<simd8x64<uint8_t>>(data, pos, len, left, right);
+  return skip_container<simd8x64<uint8_t>>(data, pos, len, left, right);
 }
 
 // TODO: optimize by removing bound checking.