@emcastillo 2024-11-06
@liuq19  sorry for the delay. I pushed some formatting changes.
I ran the files through "clang-format" hope is that enough